move atributes to config store atributue
