+++ src/client/src/App.tsx
@@ -29,7 +29,7 @@ import { MODAL_TYPES } from "./store/modals/typeKeys";
 import RightSidebar from "./containers/RightSidebar";
 import TopNavBar from "./components/TopNavBar";
 import { setPagesAction } from "./store/selection/pages/action";
-import { setValidationsAction } from "./store/selection/validations/action";
+import { setValidationsAction } from "./store/config/validations/action";
 import { setOutputPathAction } from "./store/selection/app/action";
 import { setFrontendFrameworksAction, setBackendFrameworksAction } from "./store/templates/frameworks/action";
 import { getPagesOptionsAction } from "./store/templates/pages/action";
