+++ src/client/src/store/config/combineReducers.ts
@@ -1,10 +1,14 @@
 import { combineReducers } from "redux";
 import detailsPage from "./pages/detailsPageReducer";
 import previewStatus from "./config/previewReducer";
+import validations from "./validations/setValidationsReducer";
+import isValidatingName from "./validations/validatingNameReducer";
 
 const configReducer = combineReducers({
   detailsPage,
-  previewStatus
+  previewStatus,
+  validations,
+  isValidatingName
 });
 
 export default configReducer;
