+++ src/client/src/store/selection/selectionActionType.ts
@@ -1,6 +1,6 @@
 import { ISelectFrontendAction, ISelectBackendAction } from "./frameworks/model";
 import { IsetPagesAction, ISetPageAction, IResetPagesAction } from "./pages/model";
-import { IProjectPathValidationAction, IValidationsAction } from "./validations/model";
+import { IProjectPathValidationAction, IValidationsAction } from "../config/validations/model";
 import { ISelectProjectTypeAction, ISetProjectNameActionType, IUpdateProjectPathActionType } from "./app/model";
 
 type WizardSelectionActionType =
