+++ src/client/src/utils/validations/itemName/itemName.ts
@@ -1,7 +1,7 @@
 import { addExistingItemNameValidate, addRegexValidate, addRequiredValidate, addReservedNameValidate, IValidation} from '../validations';
 import { validationMessages } from '../messages';
 import { ISelected } from "../../../types/selected";
-import { IitemNameValidationConfig } from '../../../store/selection/validations/model';
+import { IitemNameValidationConfig } from '../../../store/config/validations/model';
 
 export const validateItemName = async (itemName: string,
   validations: IitemNameValidationConfig,
