+++ src/client/src/utils/validations/validations.ts
@@ -4,7 +4,7 @@ import { FormattedMessage } from "react-intl";
 import { validationMessages } from './messages';
 
 import {projectPathValidation} from "../extensionService/extensionService";
-import { IRegex } from "../../store/selection/validations/model";
+import { IRegex } from "../../store/config/validations/model";
 
 export interface IValidation {
   isValid: boolean;
